feat: Disallow demo user from deleting base products.

+++ app/components/DeleteButton.tsx
@@ -17,14 +17,36 @@ const DeleteButton = ({ id }: { id: string }) => {
 		return;
 	}
 
+	const demoAdmin = "admin.pizza@king.com";
+	const baseProduct = [
+		"65cba0882f878312666ee9ba",
+		"65cba10f2f878312666ee9bb",
+		"65cba2ac2f878312666ee9bc",
+		"65cba3042f878312666ee9bd",
+		"65cba3862f878312666ee9be",
+		"65cc5f1dfcdc4e94e6b0abd1",
+		"65cc5fb3fcdc4e94e6b0abd2",
+		"65cc602efcdc4e94e6b0abd3",
+		"65cc798cfcdc4e94e6b0abd4",
+		"65cc7a01fcdc4e94e6b0abd5",
+		"65cc7a48fcdc4e94e6b0abd6",
+		"65cc7e04fcdc4e94e6b0abdb",
+	];
 	const handleDelete = async () => {
-		const res = await fetch(`http://localhost:3000/api/products/${id}`, {
-			method: "DELETE",
-		});
+		if (session?.user.email === demoAdmin && baseProduct.includes(id)) {
+			toast.error(
+				"Demo admins can't delete base files. Add a product to test the delete functionality."
+			);
+			return;
+		} else {
+			const res = await fetch(`${process.env.BASE_URL}/api/products/${id}`, {
+				method: "DELETE",
+			});
 
-		if (res.status === 200) {
-			router.push("/menu");
-			toast.success("The product has been deleted!");
+			if (res.status === 200) {
+				router.push("/menu");
+				toast.success("The product has been deleted!");
+			}
 		}
 	};
